Rework the way phases are stored and executed
(defphase) can now specify a 'defaults' phase to inherit from. One is
provided that has default implementations of begin, end, get-handler,
:chat, :ready, and :disconnect.

The 'default handler has been replaced with get-handler, which is given
a message and expected to return the handler for that message. The
default implementation returns a disconnect handler for messages from
non-logged-in players, the appropriate handler for messages that the
current phase understands, and a logging function for everything else.